+++ service/store/src/main/java/com/webest/store/product/application/ProductService.java
@@ -9,8 +9,8 @@ import com.webest.store.product.domain.ProductRepository;
 import com.webest.store.product.domain.ProductStatus;
 import com.webest.store.product.exception.ProductErrorCode;
 import com.webest.store.product.exception.ProductException;
-import com.webest.store.store.domain.Store;
-import com.webest.store.store.domain.StoreRepository;
+import com.webest.store.store.domain.model.Store;
+import com.webest.store.store.domain.repository.StoreRepository;
 import com.webest.store.store.exception.StoreErrorCode;
 import com.webest.store.store.exception.StoreException;
 import lombok.RequiredArgsConstructor;
