+++ service/store/src/main/java/com/webest/store/store/domain/model/Store.java
@@ -1,8 +1,6 @@
-package com.webest.store.store.domain;
+package com.webest.store.store.domain.model;
 
 import com.webest.app.jpa.BaseEntity;
-import com.webest.store.product.domain.Product;
-import com.webest.store.store.domain.converter.StoreAddressConverter;
 import jakarta.persistence.Entity;
 import jakarta.persistence.*;
 import jakarta.persistence.GeneratedValue;
@@ -58,8 +56,11 @@ public class Store extends BaseEntity {
 
     private Double deliveryTip; // 배달팁
 
-    @Column(name = "address_codes")
-    @Convert(converter = StoreAddressConverter.class)
+//    @Column(name = "address_codes")
+//    @Convert(converter = StoreAddressConverter.class)
+    @ElementCollection
+    @CollectionTable(name = "store_address_codes", joinColumns = @JoinColumn(name = "store_id"))
+    @Column(name = "address_code")
     private List<Long> addressCodeList = new ArrayList<>();
 
     public static Store of(String name, Long ownerId, Long categoryId, Integer preparationTime, Double minimumOrderAmount, String phone, LocalTime openTime, LocalTime closeTime) {
