+++ service/store/src/main/java/com/webest/store/store/domain/repository/StoreRepository.java
@@ -1,5 +1,6 @@
-package com.webest.store.store.domain;
+package com.webest.store.store.domain.repository;
 
+import com.webest.store.store.domain.model.Store;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface StoreRepository extends JpaRepository<Store, Long> {
