+++ service/store/src/main/java/com/webest/store/store/presentation/dto/StoreResponse.java
@@ -1,15 +1,8 @@
 package com.webest.store.store.presentation.dto;
 
-import com.fasterxml.jackson.databind.annotation.JsonDeserialize;
-import com.fasterxml.jackson.databind.annotation.JsonSerialize;
-import com.fasterxml.jackson.datatype.jsr310.deser.LocalDateTimeDeserializer;
-import com.fasterxml.jackson.datatype.jsr310.deser.LocalTimeDeserializer;
-import com.fasterxml.jackson.datatype.jsr310.ser.LocalDateTimeSerializer;
-import com.fasterxml.jackson.datatype.jsr310.ser.LocalTimeSerializer;
-import com.webest.store.store.domain.Store;
-import com.webest.store.store.domain.StoreStatus;
+import com.webest.store.store.domain.model.Store;
+import com.webest.store.store.domain.model.StoreStatus;
 
-import java.io.Serializable;
 import java.time.LocalTime;
 import java.util.List;
